+++ packages/paco/tests/shared.test.js
@@ -26,8 +26,7 @@ testPatternDrafting(
   Paco,
   expect,
   models,
-  patterns,
-  true
+  patterns
 )
 
 // Test sampling