+++ tests/patterns/drafting.js
@@ -1,3 +1,5 @@
+const nonHumanMeasurements = require('./non-human-measurements.js')
+
 // Some patterns are different
 const isGarment = design => ([
   'rendertest',
@@ -18,6 +20,9 @@ const isGarment = design => ([
  * @param object patterns: Imported @freesewing/pattern-info
  */
 const testPatternDrafting = (design, Pattern, expect, models, patterns, log=false) => {
+  // Load non-human measurements
+  const nonHuman = nonHumanMeasurements(models)
+
   // Helper method to try/catch pattern drafting
   const doesItDraft = (pattern, log=false) => {
     try {
@@ -41,7 +46,7 @@ const testPatternDrafting = (design, Pattern, expect, models, patterns, log=fals
    * Draft pattern for different models
    */
   if (isGarment(design)) {
-    it('Draft for human models:', () => true)
+    it('Draft for humans:', () => true)
 
     for (let size in ourModels) {
       it(`  - Drafting for ${size} (${breasts ? 'with' : 'no'} breasts)`, () => {
@@ -56,32 +61,28 @@ const testPatternDrafting = (design, Pattern, expect, models, patterns, log=fals
     }
 
     // Do the same for fantistical models (dolls, giants)
-    it('Draft for non-human models:', () => true)
+    it('Draft for dolls:', () => true)
 
-    const fractionModel = fraction => {
-      const model = {}
-      for (const [measie, value] of Object.entries(ourModels.size40)) {
-        model[measie] = value * fraction
-      }
-
-      return model
-    }
-    const models = {
-      // These are just names, don't read too much into it
-      'Gnome (0.1)': fractionModel(0.1),
-      'Halfling (0.2)': fractionModel(0.2),
-      'Goblin (0.5)': fractionModel(0.5),
-      'Dwarf (0.75)': fractionModel(0.75),
-      'Elf (1.5)': fractionModel(1.5),
-      'Ogre (2.5)': fractionModel(2.5),
-      'Firbolg (5)': fractionModel(5),
-    }
-    for (let size in models) {
+    for (let size in nonHuman[breasts ? 'withBreasts' : 'withoutBreasts'].dolls) {
       it(`  - Drafting for ${size} (${breasts ? 'with' : 'no'} breasts)`, () => {
         expect(
           doesItDraft(
             new Pattern({
-              measurements: models[size]
+              measurements: nonHuman[breasts ? 'withBreasts' : 'withoutBreasts'].dolls[size]
+            }), log
+          )
+        ).to.equal(true)
+      })
+    }
+
+    it('Draft for giants:', () => true)
+
+    for (let size in nonHuman[breasts ? 'withBreasts' : 'withoutBreasts'].giants) {
+      it(`  - Drafting for ${size} (${breasts ? 'with' : 'no'} breasts)`, () => {
+        expect(
+          doesItDraft(
+            new Pattern({
+              measurements: nonHuman[breasts ? 'withBreasts' : 'withoutBreasts'].giants[size]
             }), log
           )
         ).to.equal(true)